Streaming Record Conversion enabled
