+++ etc/source-upload.properties.template
@@ -6,6 +6,7 @@ upload.source.client.secret=upload_secret
 upload.source.client.tokenUrl=http://managementportal-app:8080/managementportal/oauth/token
 upload.source.backend.baseUrl=http://radar-upload-connect-backend:8085/radar-upload/
 upload.source.poll.interval.ms=1800000
+upload.source.queue.size=1000
 upload.source.record.converter.classes=org.radarbase.connect.upload.converter.phone.AccelerometerConverterFactory,org.radarbase.connect.upload.converter.altoida.AltoidaZipConverterFactory
 upload.source.file.uploader.username=accesskey
 upload.source.file.uploader.password=secretkey
