+++ kafka-connect-upload-source/src/main/java/org/radarbase/connect/upload/UploadSourceConnectorConfig.kt
@@ -43,7 +43,7 @@ import java.net.URI
 import java.util.concurrent.TimeUnit
 
 class UploadSourceConnectorConfig(config: ConfigDef, parsedConfig: Map<String, String>) :
-        AbstractConfig(config, parsedConfig) {
+    AbstractConfig(config, parsedConfig) {
 
     private lateinit var authenticator: Authenticator
 
@@ -51,10 +51,10 @@ class UploadSourceConnectorConfig(config: ConfigDef, parsedConfig: Map<String, S
         if(!::authenticator.isInitialized) {
             logger.info("Initializing authenticator")
             authenticator = ClientCredentialsAuthorizer(
-                    httpClient,
-                    oauthClientId,
-                    oauthClientSecret,
-                    tokenRequestUrl)
+                httpClient,
+                oauthClientId,
+                oauthClientSecret,
+                tokenRequestUrl)
         }
         return authenticator
     }
@@ -74,12 +74,12 @@ class UploadSourceConnectorConfig(config: ConfigDef, parsedConfig: Map<String, S
     val fileUploaderType: UploadType? = UploadType.valueOf(getString(UPLOAD_FILE_UPLOADER_TYPE_CONFIG).toUpperCase())
 
     val fileUploadConfig: FileUploaderFactory.FileUploaderConfig = FileUploaderFactory.FileUploaderConfig(
-            targetEndpoint = getString(UPLOAD_FILE_UPLOADER_TARGET_ENDPOINT_URL_CONFIG),
-            targetRoot = getString(UPLOAD_FILE_UPLOADER_TARGET_ROOT_DIRECTORY_CONFIG),
-            username = getString(UPLOAD_FILE_UPLOADER_USERNAME_CONFIG),
-            password = getPassword(UPLOAD_FILE_UPLOADER_PASSWORD_CONFIG)?.value(),
-            sshPrivateKey = getString(UPLOAD_FILE_UPLOADER_SSH_PRIVATE_KEY_FILE_CONFIG),
-            sshPassPhrase = getPassword(UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_CONFIG)?.value()
+        targetEndpoint = getString(UPLOAD_FILE_UPLOADER_TARGET_ENDPOINT_URL_CONFIG),
+        targetRoot = getString(UPLOAD_FILE_UPLOADER_TARGET_ROOT_DIRECTORY_CONFIG),
+        username = getString(UPLOAD_FILE_UPLOADER_USERNAME_CONFIG),
+        password = getPassword(UPLOAD_FILE_UPLOADER_PASSWORD_CONFIG)?.value(),
+        sshPrivateKey = getString(UPLOAD_FILE_UPLOADER_SSH_PRIVATE_KEY_FILE_CONFIG),
+        sshPassPhrase = getPassword(UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_CONFIG)?.value()
     )
 
     val httpClient: OkHttpClient
@@ -112,17 +112,23 @@ class UploadSourceConnectorConfig(config: ConfigDef, parsedConfig: Map<String, S
         private const val UPLOAD_SOURCE_CONVERTERS_DOC = "List record converter classes that are in class-path"
         private const val UPLOAD_SOURCE_CONVERTERS_DISPLAY = "List of record converter factory class"
         private val UPLOAD_SOURCE_CONVERTERS_DEFAULT: List<String> = listOf(
-                AccelerometerConverterFactory::class.java.name,
-                AcceleratometerZipConverterFactory::class.java.name,
-                AltoidaConverterFactory::class.java.name,
-                WearableCameraConverterFactory::class.java.name,
-                AxivityConverterFactory::class.java.name,
-                Physilog5ConverterFactory::class.java.name)
+            AccelerometerConverterFactory::class.java.name,
+            AcceleratometerZipConverterFactory::class.java.name,
+            AltoidaConverterFactory::class.java.name,
+            WearableCameraConverterFactory::class.java.name,
+            AxivityConverterFactory::class.java.name,
+            Physilog5ConverterFactory::class.java.name)
 
         const val SOURCE_POLL_INTERVAL_CONFIG = "upload.source.poll.interval.ms"
         private const val SOURCE_POLL_INTERVAL_DOC = "How often to poll the source URL."
         private const val SOURCE_POLL_INTERVAL_DISPLAY = "Polling interval"
         private const val SOURCE_POLL_INTERVAL_DEFAULT = 60000L
+
+
+        const val SOURCE_QUEUE_SIZE_CONFIG = "upload.source.queue.size"
+        private const val SOURCE_QUEUE_SIZE_DOC = "Capacity of the records queue."
+        private const val SOURCE_QUEUE_SIZE_DISPLAY = "Records queue size"
+        private const val SOURCE_QUEUE_SIZE_DEFAULT = 1000
 
         private const val UPLOAD_FILE_UPLOADER_TYPE_CONFIG = "upload.source.file.uploader.type"
         private const val UPLOAD_FILE_UPLOADER_TYPE_DOC = "Choose which type of file uploader should be used to upload files to target location from local, sftp, s3."
@@ -160,152 +166,162 @@ class UploadSourceConnectorConfig(config: ConfigDef, parsedConfig: Map<String, S
         private val UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_DEFAULT: String? = null
 
         var mapper: ObjectMapper = ObjectMapper()
-                .registerModule(KotlinModule())
-                .registerModule(JavaTimeModule())
-                .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
-                .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
+            .registerModule(KotlinModule())
+            .registerModule(JavaTimeModule())
+            .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
+            .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
 
         fun conf(): ConfigDef {
             val groupName = "upload"
             var orderInGroup = 0
 
             return ConfigDef()
-                    .define(SOURCE_POLL_INTERVAL_CONFIG,
-                            ConfigDef.Type.LONG,
-                            SOURCE_POLL_INTERVAL_DEFAULT,
-                            ConfigDef.Importance.HIGH,
-                            SOURCE_POLL_INTERVAL_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.SHORT,
-                            SOURCE_POLL_INTERVAL_DISPLAY)
+                .define(SOURCE_POLL_INTERVAL_CONFIG,
+                    ConfigDef.Type.LONG,
+                    SOURCE_POLL_INTERVAL_DEFAULT,
+                    ConfigDef.Importance.HIGH,
+                    SOURCE_POLL_INTERVAL_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.SHORT,
+                    SOURCE_POLL_INTERVAL_DISPLAY)
 
-                    .define(UPLOAD_SOURCE_CLIENT_CONFIG,
-                            ConfigDef.Type.STRING,
-                            UPLOAD_SOURCE_MP_CLIENT_DEFAULT,
-                            ConfigDef.Importance.HIGH,
-                            UPLOAD_SOURCE_CLIENT_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.SHORT,
-                            UPLOAD_SOURCE_MP_CLIENT_DISPLAY)
+                .define(SOURCE_QUEUE_SIZE_CONFIG,
+                    ConfigDef.Type.INT,
+                    SOURCE_QUEUE_SIZE_DEFAULT,
+                    ConfigDef.Importance.HIGH,
+                    SOURCE_QUEUE_SIZE_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.SHORT,
+                    SOURCE_QUEUE_SIZE_DISPLAY)
 
-                    .define(UPLOAD_SOURCE_MP_SECRET_CONFIG,
-                            ConfigDef.Type.STRING,
-                            "",
-                            ConfigDef.Importance.HIGH,
-                            UPLOAD_SOURCE_MP_SECRET_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.SHORT,
-                            UPLOAD_SOURCE_MP_SECRET_DISPLAY)
+                .define(UPLOAD_SOURCE_CLIENT_CONFIG,
+                    ConfigDef.Type.STRING,
+                    UPLOAD_SOURCE_MP_CLIENT_DEFAULT,
+                    ConfigDef.Importance.HIGH,
+                    UPLOAD_SOURCE_CLIENT_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.SHORT,
+                    UPLOAD_SOURCE_MP_CLIENT_DISPLAY)
 
-                    .define(UPLOAD_SOURCE_SERVER_BASE_URL_CONFIG,
-                            ConfigDef.Type.STRING,
-                            UPLOAD_SOURCE_SERVER_BASE_URL_DEFAULT,
-                            ConfigDef.Importance.LOW,
-                            UPLOAD_SOURCE_SERVER_BASE_URL_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.SHORT,
-                            UPLOAD_SOURCE_SERVER_BASE_URL_DISPLAY)
+                .define(UPLOAD_SOURCE_MP_SECRET_CONFIG,
+                    ConfigDef.Type.STRING,
+                    "",
+                    ConfigDef.Importance.HIGH,
+                    UPLOAD_SOURCE_MP_SECRET_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.SHORT,
+                    UPLOAD_SOURCE_MP_SECRET_DISPLAY)
 
-                    .define(UPLOAD_SOURCE_CLIENT_TOKEN_URL_CONFIG,
-                            ConfigDef.Type.STRING,
-                            UPLOAD_SOURCE_CLIENT_TOKEN_URL_DEFAULT,
-                            ConfigDef.Importance.LOW,
-                            UPLOAD_SOURCE_CLIENT_TOKEN_URL_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.SHORT,
-                            UPLOAD_SOURCE_CLIENT_TOKEN_URL_DISPLAY)
+                .define(UPLOAD_SOURCE_SERVER_BASE_URL_CONFIG,
+                    ConfigDef.Type.STRING,
+                    UPLOAD_SOURCE_SERVER_BASE_URL_DEFAULT,
+                    ConfigDef.Importance.LOW,
+                    UPLOAD_SOURCE_SERVER_BASE_URL_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.SHORT,
+                    UPLOAD_SOURCE_SERVER_BASE_URL_DISPLAY)
 
-                    .define(UPLOAD_SOURCE_CONVERTERS_CONFIG,
-                            ConfigDef.Type.LIST,
-                            UPLOAD_SOURCE_CONVERTERS_DEFAULT,
-                            ConfigDef.Importance.HIGH,
-                            UPLOAD_SOURCE_CONVERTERS_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.LONG,
-                            UPLOAD_SOURCE_CONVERTERS_DISPLAY)
+                .define(UPLOAD_SOURCE_CLIENT_TOKEN_URL_CONFIG,
+                    ConfigDef.Type.STRING,
+                    UPLOAD_SOURCE_CLIENT_TOKEN_URL_DEFAULT,
+                    ConfigDef.Importance.LOW,
+                    UPLOAD_SOURCE_CLIENT_TOKEN_URL_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.SHORT,
+                    UPLOAD_SOURCE_CLIENT_TOKEN_URL_DISPLAY)
 
-                    .define(UPLOAD_FILE_UPLOADER_TYPE_CONFIG,
-                            ConfigDef.Type.STRING,
-                            UPLOAD_FILE_UPLOADER_TYPE_DEFAULT,
-                            ConfigDef.Importance.LOW,
-                            UPLOAD_FILE_UPLOADER_TYPE_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.LONG,
-                            UPLOAD_FILE_UPLOADER_TYPE_DISPLAY)
+                .define(UPLOAD_SOURCE_CONVERTERS_CONFIG,
+                    ConfigDef.Type.LIST,
+                    UPLOAD_SOURCE_CONVERTERS_DEFAULT,
+                    ConfigDef.Importance.HIGH,
+                    UPLOAD_SOURCE_CONVERTERS_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.LONG,
+                    UPLOAD_SOURCE_CONVERTERS_DISPLAY)
 
-                    .define(UPLOAD_FILE_UPLOADER_TARGET_ENDPOINT_URL_CONFIG,
-                            ConfigDef.Type.STRING,
-                            UPLOAD_FILE_UPLOADER_TARGET_ENDPOINT_URL_DEFAULT,
-                            ConfigDef.Importance.LOW,
-                            UPLOAD_FILE_UPLOADER_TARGET_ENDPOINT_URL_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.MEDIUM,
-                            UPLOAD_FILE_UPLOADER_TARGET_ENDPOINT_URL_DISPLAY)
+                .define(UPLOAD_FILE_UPLOADER_TYPE_CONFIG,
+                    ConfigDef.Type.STRING,
+                    UPLOAD_FILE_UPLOADER_TYPE_DEFAULT,
+                    ConfigDef.Importance.LOW,
+                    UPLOAD_FILE_UPLOADER_TYPE_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.LONG,
+                    UPLOAD_FILE_UPLOADER_TYPE_DISPLAY)
 
-                    .define(UPLOAD_FILE_UPLOADER_TARGET_ROOT_DIRECTORY_CONFIG,
-                            ConfigDef.Type.STRING,
-                            UPLOAD_FILE_UPLOADER_TARGET_ROOT_DIRECTORY_DEFAULT,
-                            ConfigDef.Importance.LOW,
-                            UPLOAD_FILE_UPLOADER_TARGET_ROOT_DIRECTORY_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.SHORT,
-                            UPLOAD_FILE_UPLOADER_TARGET_ROOT_DIRECTORY_DISPLAY)
+                .define(UPLOAD_FILE_UPLOADER_TARGET_ENDPOINT_URL_CONFIG,
+                    ConfigDef.Type.STRING,
+                    UPLOAD_FILE_UPLOADER_TARGET_ENDPOINT_URL_DEFAULT,
+                    ConfigDef.Importance.LOW,
+                    UPLOAD_FILE_UPLOADER_TARGET_ENDPOINT_URL_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.MEDIUM,
+                    UPLOAD_FILE_UPLOADER_TARGET_ENDPOINT_URL_DISPLAY)
 
-                    .define(UPLOAD_FILE_UPLOADER_USERNAME_CONFIG,
-                            ConfigDef.Type.STRING,
-                            UPLOAD_FILE_UPLOADER_USERNAME_DEFAULT,
-                            ConfigDef.Importance.LOW,
-                            UPLOAD_FILE_UPLOADER_USERNAME_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.MEDIUM,
-                            UPLOAD_FILE_UPLOADER_USERNAME_DISPLAY)
+                .define(UPLOAD_FILE_UPLOADER_TARGET_ROOT_DIRECTORY_CONFIG,
+                    ConfigDef.Type.STRING,
+                    UPLOAD_FILE_UPLOADER_TARGET_ROOT_DIRECTORY_DEFAULT,
+                    ConfigDef.Importance.LOW,
+                    UPLOAD_FILE_UPLOADER_TARGET_ROOT_DIRECTORY_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.SHORT,
+                    UPLOAD_FILE_UPLOADER_TARGET_ROOT_DIRECTORY_DISPLAY)
 
-                    .define(UPLOAD_FILE_UPLOADER_PASSWORD_CONFIG,
-                            ConfigDef.Type.PASSWORD,
-                            UPLOAD_FILE_UPLOADER_PASSWORD_DEFAULT,
-                            ConfigDef.Importance.LOW,
-                            UPLOAD_FILE_UPLOADER_PASSWORD_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.MEDIUM,
-                            UPLOAD_FILE_UPLOADER_PASSWORD_DISPLAY)
+                .define(UPLOAD_FILE_UPLOADER_USERNAME_CONFIG,
+                    ConfigDef.Type.STRING,
+                    UPLOAD_FILE_UPLOADER_USERNAME_DEFAULT,
+                    ConfigDef.Importance.LOW,
+                    UPLOAD_FILE_UPLOADER_USERNAME_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.MEDIUM,
+                    UPLOAD_FILE_UPLOADER_USERNAME_DISPLAY)
 
-                    .define(UPLOAD_FILE_UPLOADER_SSH_PRIVATE_KEY_FILE_CONFIG,
-                            ConfigDef.Type.STRING,
-                            UPLOAD_FILE_UPLOADER_SSH_PRIVATE_KEY_FILE_DEFAULT,
-                            ConfigDef.Importance.LOW,
-                            UPLOAD_FILE_UPLOADER_SSH_PRIVATE_KEY_FILE_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.MEDIUM,
-                            UPLOAD_FILE_UPLOADER_SSH_PRIVATE_KEY_FILE_DISPLAY)
+                .define(UPLOAD_FILE_UPLOADER_PASSWORD_CONFIG,
+                    ConfigDef.Type.PASSWORD,
+                    UPLOAD_FILE_UPLOADER_PASSWORD_DEFAULT,
+                    ConfigDef.Importance.LOW,
+                    UPLOAD_FILE_UPLOADER_PASSWORD_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.MEDIUM,
+                    UPLOAD_FILE_UPLOADER_PASSWORD_DISPLAY)
 
-                    .define(UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_CONFIG,
-                            ConfigDef.Type.PASSWORD,
-                            UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_DEFAULT,
-                            ConfigDef.Importance.LOW,
-                            UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_DOC,
-                            groupName,
-                            ++orderInGroup,
-                            ConfigDef.Width.MEDIUM,
-                            UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_DISPLAY)
+                .define(UPLOAD_FILE_UPLOADER_SSH_PRIVATE_KEY_FILE_CONFIG,
+                    ConfigDef.Type.STRING,
+                    UPLOAD_FILE_UPLOADER_SSH_PRIVATE_KEY_FILE_DEFAULT,
+                    ConfigDef.Importance.LOW,
+                    UPLOAD_FILE_UPLOADER_SSH_PRIVATE_KEY_FILE_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.MEDIUM,
+                    UPLOAD_FILE_UPLOADER_SSH_PRIVATE_KEY_FILE_DISPLAY)
+
+                .define(UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_CONFIG,
+                    ConfigDef.Type.PASSWORD,
+                    UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_DEFAULT,
+                    ConfigDef.Importance.LOW,
+                    UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_DOC,
+                    groupName,
+                    ++orderInGroup,
+                    ConfigDef.Width.MEDIUM,
+                    UPLOAD_FILE_UPLOADER_SSH_PASSPHRASE_DISPLAY)
 
         }
 
         private val globalHttpClient = OkHttpClient.Builder()
-                .connectTimeout(10, TimeUnit.SECONDS)
-                .writeTimeout(10, TimeUnit.SECONDS)
-                .readTimeout(30, TimeUnit.SECONDS)
-                .build()
+            .connectTimeout(10, TimeUnit.SECONDS)
+            .writeTimeout(10, TimeUnit.SECONDS)
+            .readTimeout(30, TimeUnit.SECONDS)
+            .build()
     }
 }
