+++ kafka-connect-upload-source/src/main/java/org/radarbase/connect/upload/converter/ConverterFactory.kt
@@ -80,7 +80,7 @@ interface ConverterFactory {
         val sourceType: String
 
         // convert and add logs return result
-        fun convert(record: RecordDTO): List<SourceRecord>
+        fun convert(record: RecordDTO): Sequence<SourceRecord>
 
         fun convertFile(record: RecordDTO, contents: ContentsDTO, inputStream: InputStream, recordLogger: RecordLogger): List<TopicData>
