+++ kafka-connect-upload-source/src/test/java/org/radarbase/connect/upload/UploadSourceConnectorConfigTest.kt
@@ -37,6 +37,7 @@ class UploadSourceConnectorConfigTest {
                 "upload.source.backend.baseUrl" to
                         "http://radar-upload-connect-backend:8085/radar-upload/",
                 "upload.source.poll.interval.ms" to "3600000",
+                "upload.source.queue.size" to "1000",
                 "upload.source.record.converter.classes" to listOf(
                         AccelerometerConverterFactory::class.java.name,
                         AltoidaConverterFactory::class.java.name
